feat(webhook): common sender class
